# AST: Optimize removeShadowedDecls()