@slavapestov 2020-05-27
@swift-ci Please test compiler performance