@slavapestov 2020-05-28
# Summary for master full

**Unexpected test results, excluded stats for RxCocoa, Base64CoderSwiftUI, SwifterSwift, Backend**

**Regressions found (see below)**

- [Debug-batch](#debug-batch-7297af0c54a8c984)
    - [debug-batch brief](#debug-batch-brief-7297af0c54a8c984)
    - [debug-batch detailed](#debug-batch-detailed-7297af0c54a8c984)
- [Release](#release-7297af0c54a8c984)
    - [release brief](#release-brief-7297af0c54a8c984)
    - [release detailed](#release-detailed-7297af0c54a8c984)



<a name="debug-batch-7297af0c54a8c984">

# Debug-batch

<a name="debug-batch-brief-7297af0c54a8c984">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 64,541,600,134,354 | 60,260,573,091,382 | -4,281,027,042,972 | -6.63% :white_check_mark:
time.swift-driver.wall | 5556.5s | 5410.1s | -146.4s | -2.63% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 1,872,966,460 | 1,872,968,170 | 1,710 | 0.0%
</details>

<a name="debug-batch-detailed-7297af0c54a8c984">

## debug-batch detailed
<details>
<summary>Regressed (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumSourceLinesPerSecond | 2,640,475 | 2,876,106 | 235,631 | 8.92% :no_entry:
</details>
<details>
<summary>Improved (15)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumASTBytesAllocated | 63,748,883,120 | 61,874,877,664 | -1,874,005,456 | -2.94% :white_check_mark:
Driver.ChildrenMaxRSS | 222,637,359,104 | 216,944,807,936 | -5,692,551,168 | -2.56% :white_check_mark:
Driver.NumDriverPipePolls | 87,264 | 79,492 | -7,772 | -8.91% :white_check_mark:
Driver.NumDriverPipeReads | 76,256 | 68,553 | -7,703 | -10.1% :white_check_mark:
Frontend.MaxMallocUsage | 1,135,741,317,752 | 1,086,330,110,848 | -49,411,206,904 | -4.35% :white_check_mark:
Frontend.NumInstructionsExecuted | 64,541,600,134,354 | 60,260,573,091,382 | -4,281,027,042,972 | -6.63% :white_check_mark:
Sema.GenericSignatureRequest | 2,606,163 | 2,541,330 | -64,833 | -2.49% :white_check_mark:
Sema.LookupConformanceInModuleRequest | 38,313,578 | 37,336,950 | -976,628 | -2.55% :white_check_mark:
Sema.NumConformancesDeserialized | 10,354,026 | 10,244,193 | -109,833 | -1.06% :white_check_mark:
Sema.NumGenericSignatureBuilders | 1,445,908 | 1,258,704 | -187,204 | -12.95% :white_check_mark:
Sema.NumLazyIterableDeclContexts | 9,439,163 | 9,322,238 | -116,925 | -1.24% :white_check_mark:
Sema.NumLazyRequirementSignaturesLoaded | 670,934 | 610,928 | -60,006 | -8.94% :white_check_mark:
Sema.NumTypesDeserialized | 25,115,262 | 24,848,017 | -267,245 | -1.06% :white_check_mark:
Sema.NumUnloadedLazyIterableDeclContexts | 5,864,343 | 5,746,908 | -117,435 | -2.0% :white_check_mark:
Sema.RequirementSignatureRequest | 731,032 | 671,026 | -60,006 | -8.21% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (224)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 2,004,963 | 2,005,257 | 294 | 0.01%
AST.ImportSetCacheMiss | 463,948 | 463,992 | 44 | 0.01%
AST.ImportSetFoldHit | 144,736 | 144,774 | 38 | 0.03%
AST.ImportSetFoldMiss | 319,212 | 319,218 | 6 | 0.0%
AST.ModuleShadowCacheHit | 3,715 | 3,715 | 0 | 0.0%
AST.ModuleShadowCacheMiss | 1,956 | 1,956 | 0 | 0.0%
AST.ModuleVisibilityCacheHit | 38,909 | 38,909 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 9,291 | 9,291 | 0 | 0.0%
AST.NumASTScopeLookups | 3,976,324 | 3,976,328 | 4 | 0.0%
AST.NumBraceStmtASTScopeExpansions | 646,560 | 646,560 | 0 | 0.0%
AST.NumBraceStmtASTScopes | 646,560 | 646,560 | 0 | 0.0%
AST.NumDecls | 146,140 | 146,140 | 0 | 0.0%
AST.NumDependencies | 362,843 | 362,857 | 14 | 0.0%
AST.NumIterableTypeBodyASTScopeExpansions | 261,520 | 261,523 | 3 | 0.0%
AST.NumIterableTypeBodyASTScopes | 325,724 | 325,465 | -259 | -0.08%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 293,688 | 293,688 | 0 | 0.0%
AST.NumLocalTypeDecls | 255 | 255 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,993 | 6,993 | 0 | 0.0%
AST.NumModuleLookupValue | 52,153,439 | 52,154,114 | 675 | 0.0%
AST.NumObjCMethods | 25,372 | 25,372 | 0 | 0.0%
AST.NumOperators | 589 | 589 | 0 | 0.0%
AST.NumPrecedenceGroups | 93 | 93 | 0 | 0.0%
AST.NumReferencedDynamicNames | 191 | 191 | 0 | 0.0%
AST.NumReferencedMemberNames | 7,731,448 | 7,731,274 | -174 | -0.0%
AST.NumReferencedTopLevelNames | 615,392 | 615,392 | 0 | 0.0%
AST.NumSourceBuffers | 423,778 | 423,778 | 0 | 0.0%
AST.NumSourceLines | 4,861,186 | 4,861,186 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 2,113,877 | 2,115,349 | 1,472 | 0.07%
Driver.DriverDepCascadingDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingTopLevel | 0 | 0 | 0 | 0.0%
Driver.DriverDepDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepTopLevel | 0 | 0 | 0 | 0.0%
Driver.NumDriverJobsRun | 29,026 | 29,026 | 0 | 0.0%
Driver.NumDriverJobsSkipped | 0 | 0 | 0 | 0.0%
Driver.NumProcessFailures | 0 | 0 | 0 | 0.0%
Frontend.NumProcessFailures | 0 | 0 | 0 | 0.0%
IRGen.IRGenSourceFileRequest | 27,711 | 27,711 | 0 | 0.0%
IRGen.IRGenWholeModuleRequest | 20 | 20 | 0 | 0.0%
IRModule.NumGOTEntries | 249,591 | 249,591 | 0 | 0.0%
IRModule.NumIRAliases | 203,760 | 203,760 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 7,062,889 | 7,062,889 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 3,553,274 | 3,553,274 | 0 | 0.0%
IRModule.NumIRGlobals | 3,853,891 | 3,853,891 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 90,772,176 | 90,772,176 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 140,155 | 140,155 | 0 | 0.0%
IRModule.NumIRValueSymbols | 6,728,546 | 6,728,546 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 1,872,966,460 | 1,872,968,170 | 1,710 | 0.0%
Parse.CodeCompletionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 280,457 | 280,457 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 758,609 | 758,619 | 10 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 258,079 | 258,079 | 0 | 0.0%
Parse.ParseMembersRequest | 611,885 | 611,895 | 10 | 0.0%
Parse.ParseSourceFileRequest | 413,955 | 413,955 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILGen.SILGenerationRequest | 29,014 | 29,014 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 1,773,652 | 1,773,652 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 55,396 | 55,396 | 0 | 0.0%
SILModule.NumSILGenVtables | 19,143 | 19,143 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 75,979 | 75,979 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 2,613,444 | 2,613,444 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 57,115 | 57,115 | 0 | 0.0%
SILModule.NumSILOptVtables | 32,441 | 32,441 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 175,886 | 175,886 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 112,207 | 112,207 | 0 | 0.0%
Sema.AbstractGenericSignatureRequest | 36,988 | 36,988 | 0 | 0.0%
Sema.AccessLevelRequest | 14,543,034 | 14,544,765 | 1,731 | 0.01%
Sema.AnyObjectLookupRequest | 283 | 283 | 0 | 0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 23,023 | 23,023 | 0 | 0.0%
Sema.AttachedFunctionBuilderRequest | 351,518 | 351,518 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 588,084 | 588,084 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 2,654,617 | 2,654,620 | 3 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 79,508 | 79,508 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,527 | 10,527 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 1,015,032 | 1,015,032 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 109,230 | 109,230 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 93,523 | 93,523 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 9,270,595 | 9,277,822 | 7,227 | 0.08%
Sema.CompareDeclSpecializationRequest | 500,271 | 500,772 | 501 | 0.1%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 504 | 504 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 57,908 | 57,908 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 39,942 | 39,942 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 259 | 259 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 8,535 | 8,535 | 0 | 0.0%
Sema.DefaultTypeRequest | 500,588 | 500,588 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 35,699,447 | 35,701,914 | 2,467 | 0.01%
Sema.DirectOperatorLookupRequest | 1,902,620 | 1,902,828 | 208 | 0.01%
Sema.DirectPrecedenceGroupLookupRequest | 615,800 | 615,800 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 1,075,695 | 1,075,695 | 0 | 0.0%
Sema.EmittedMembersRequest | 27,630 | 27,630 | 0 | 0.0%
Sema.EnumRawTypeRequest | 24,671 | 24,671 | 0 | 0.0%
Sema.EnumRawValuesRequest | 10,922 | 10,922 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 12,892 | 12,915 | 23 | 0.18%
Sema.ExistentialTypeSupportedRequest | 17,332 | 17,332 | 0 | 0.0%
Sema.ExpandASTScopeRequest | 7,292,182 | 7,291,734 | -448 | -0.01%
Sema.ExtendedNominalRequest | 653,407 | 653,407 | 0 | 0.0%
Sema.ExtendedTypeRequest | 82,813 | 82,815 | 2 | 0.0%
Sema.FragileFunctionKindRequest | 1,866,060 | 1,865,939 | -121 | -0.01%
Sema.FunctionBuilderTypeRequest | 351,178 | 351,178 | 0 | 0.0%
Sema.FunctionOperatorRequest | 60,566 | 60,566 | 0 | 0.0%
Sema.GenericParamListRequest | 10,640,457 | 10,639,438 | -1,019 | -0.01%
Sema.GetDestructorRequest | 28,223 | 28,223 | 0 | 0.0%
Sema.HasCircularInheritanceRequest | 23,571 | 23,571 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 10,529 | 10,529 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 9,295 | 9,295 | 0 | 0.0%
Sema.HasDefaultInitRequest | 56,151 | 56,151 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 0 | 0 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 633,439 | 633,469 | 30 | 0.0%
Sema.HasImplementationOnlyImportsRequest | 413,955 | 413,955 | 0 | 0.0%
Sema.HasMemberwiseInitRequest | 21,139 | 21,139 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 26,154 | 26,154 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 56,161 | 56,161 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 175,416 | 175,421 | 5 | 0.0%
Sema.InheritedDeclsReferencedRequest | 5,958,433 | 5,956,184 | -2,249 | -0.04%
Sema.InheritedProtocolsRequest | 600,337 | 600,386 | 49 | 0.01%
Sema.InheritedTypeRequest | 302,692 | 302,607 | -85 | -0.03%
Sema.InheritsSuperclassInitializersRequest | 27,657 | 27,657 | 0 | 0.0%
Sema.InitKindRequest | 97,936 | 97,936 | 0 | 0.0%
Sema.InterfaceTypeRequest | 15,487,341 | 15,473,622 | -13,719 | -0.09%
Sema.IsABICompatibleOverrideRequest | 140,761 | 140,761 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 325,763 | 325,763 | 0 | 0.0%
Sema.IsCallableNominalTypeRequest | 2,173 | 2,173 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDynamicRequest | 1,781,333 | 1,781,333 | 0 | 0.0%
Sema.IsFinalRequest | 2,912,763 | 2,914,651 | 1,888 | 0.06%
Sema.IsGetterMutatingRequest | 443,867 | 443,867 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 2,482,110 | 2,482,277 | 167 | 0.01%
Sema.IsObjCRequest | 1,618,968 | 1,618,590 | -378 | -0.02%
Sema.IsSetterMutatingRequest | 381,246 | 381,246 | 0 | 0.0%
Sema.IsStaticRequest | 938,890 | 938,939 | 49 | 0.01%
Sema.LazyStoragePropertyRequest | 2,745 | 2,745 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 85,189 | 85,189 | 0 | 0.0%
Sema.LookupInModuleRequest | 6,983,099 | 6,982,208 | -891 | -0.01%
Sema.LookupInfixOperatorRequest | 85,912 | 85,912 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 186 | 186 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 27,812 | 27,812 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 598 | 598 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 510 | 510 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 11,810 | 11,810 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 3,278,676 | 3,277,784 | -892 | -0.03%
Sema.NamedLazyMemberLoadSuccessCount | 21,819,720 | 21,827,397 | 7,677 | 0.04%
Sema.NamingPatternRequest | 230,586 | 230,589 | 3 | 0.0%
Sema.NeedsNewVTableEntryRequest | 706,166 | 706,166 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 196,151 | 196,151 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 301,349 | 301,349 | 0 | 0.0%
Sema.NumConstraintScopes | 31,401,727 | 31,402,626 | 899 | 0.0%
Sema.NumConstraintsConsideredForEdgeContraction | 103,438,722 | 103,439,135 | 413 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 269 | 269 | 0 | 0.0%
Sema.NumDeclsDeserialized | 81,245,748 | 80,895,733 | -350,015 | -0.43%
Sema.NumDeclsTypechecked | 1,463,819 | 1,463,819 | 0 | 0.0%
Sema.NumLazyRequirementSignatures | 918,049 | 909,725 | -8,324 | -0.91%
Sema.NumLeafScopes | 20,048,144 | 20,048,752 | 608 | 0.0%
Sema.NumTypesValidated | 667,325 | 667,328 | 3 | 0.0%
Sema.OpaqueReadOwnershipRequest | 284,581 | 284,581 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 197 | 197 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 749 | 749 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 2,758,711 | 2,759,566 | 855 | 0.03%
Sema.ParamSpecifierRequest | 1,278,763 | 1,278,709 | -54 | -0.0%
Sema.PatternBindingEntryRequest | 489,882 | 489,885 | 3 | 0.0%
Sema.PatternTypeRequest | 565,648 | 565,651 | 3 | 0.0%
Sema.PreCheckFunctionBuilderRequest | 312 | 312 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyInfoRequest | 581,783 | 581,783 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 588,084 | 588,084 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 250 | 250 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 629,737 | 629,737 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 96 | 96 | 0 | 0.0%
Sema.ProtocolRequiresClassRequest | 40,734 | 40,812 | 78 | 0.19%
Sema.ProvideDefaultImplForRequest | 9,270,595 | 9,277,822 | 7,227 | 0.08%
Sema.QualifiedLookupRequest | 5,816,354 | 5,816,344 | -10 | -0.0%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RequirementRequest | 116,600 | 116,601 | 1 | 0.0%
Sema.RequiresOpaqueAccessorsRequest | 1,357,079 | 1,357,079 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 274,434 | 274,434 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 402,654 | 402,557 | -97 | -0.02%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 3,538,991 | 3,538,667 | -324 | -0.01%
Sema.ResultTypeRequest | 682,662 | 682,623 | -39 | -0.01%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 16,541,350 | 16,546,946 | 5,596 | 0.03%
Sema.ScopedImportLookupRequest | 1,076 | 1,076 | 0 | 0.0%
Sema.SelfAccessKindRequest | 649,961 | 649,869 | -92 | -0.01%
Sema.SelfBoundsFromWhereClauseRequest | 1,790,333 | 1,792,472 | 2,139 | 0.12%
Sema.SetterAccessLevelRequest | 150,819 | 150,819 | 0 | 0.0%
Sema.SimpleDidSetRequest | 2,896,169 | 2,897,437 | 1,268 | 0.04%
Sema.StorageImplInfoRequest | 1,506,656 | 1,506,656 | 0 | 0.0%
Sema.StoredPropertiesAndMissingMembersRequest | 33,225 | 33,225 | 0 | 0.0%
Sema.StoredPropertiesRequest | 357,095 | 357,095 | 0 | 0.0%
Sema.StructuralTypeRequest | 2,155 | 2,155 | 0 | 0.0%
Sema.SuperclassDeclRequest | 466,017 | 466,124 | 107 | 0.02%
Sema.SuperclassTypeRequest | 57,984 | 57,984 | 0 | 0.0%
Sema.SynthesizeAccessorRequest | 301,349 | 301,349 | 0 | 0.0%
Sema.SynthesizeDefaultInitRequest | 5,259 | 5,259 | 0 | 0.0%
Sema.SynthesizeMemberwiseInitRequest | 3,670 | 3,670 | 0 | 0.0%
Sema.TypeCheckFunctionBodyUntilRequest | 548,663 | 548,663 | 0 | 0.0%
Sema.TypeCheckSourceFileRequest | 28,031 | 28,031 | 0 | 0.0%
Sema.TypeDeclsFromWhereClauseRequest | 30,733 | 30,733 | 0 | 0.0%
Sema.TypeEraserHasViableInitRequest | 0 | 0 | 0 | 0.0%
Sema.TypeRelationCheckRequest | 0 | 0 | 0 | 0.0%
Sema.TypeWitnessRequest | 9,624 | 9,625 | 1 | 0.01%
Sema.USRGenerationRequest | 10,756,759 | 10,765,555 | 8,796 | 0.08%
Sema.UnderlyingTypeDeclsReferencedRequest | 298,001 | 298,298 | 297 | 0.1%
Sema.UnderlyingTypeRequest | 37,207 | 37,207 | 0 | 0.0%
Sema.UnqualifiedLookupRequest | 4,127,616 | 4,127,620 | 4 | 0.0%
Sema.ValidatePrecedenceGroupRequest | 183,973 | 183,973 | 0 | 0.0%
Sema.ValueWitnessRequest | 68,839 | 68,839 | 0 | 0.0%
TBDGen.GenerateTBDRequest | 27,731 | 27,731 | 0 | 0.0%
</details>

<a name="release-7297af0c54a8c984">

# Release

<a name="release-brief-7297af0c54a8c984">

## release brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 55,760,702,980,005 | 54,957,991,439,054 | -802,711,540,951 | -1.44% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 1,889,034,442 | 1,889,038,252 | 3,810 | 0.0%
time.swift-driver.wall | 9044.6s | 8955.2s | -89.4s | -0.99%
</details>

<a name="release-detailed-7297af0c54a8c984">

## release detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Sema.NumGenericSignatureBuilders | 280,337 | 237,702 | -42,635 | -15.21% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (16)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumLoadedModules | 27,738 | 27,738 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 478,300 | 478,300 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 5,012,952 | 5,012,952 | 0 | 0.0%
IRModule.NumIRFunctions | 2,895,989 | 2,895,989 | 0 | 0.0%
IRModule.NumIRGlobals | 3,408,494 | 3,408,494 | 0 | 0.0%
IRModule.NumIRInsts | 50,584,388 | 50,584,388 | 0 | 0.0%
IRModule.NumIRValueSymbols | 5,933,503 | 5,933,503 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 1,889,034,442 | 1,889,038,252 | 3,810 | 0.0%
SILModule.NumSILGenFunctions | 1,243,573 | 1,243,573 | 0 | 0.0%
SILModule.NumSILOptFunctions | 1,041,679 | 1,041,679 | 0 | 0.0%
Sema.NumConformancesDeserialized | 3,596,883 | 3,588,177 | -8,706 | -0.24%
Sema.NumConstraintScopes | 31,700,283 | 31,700,304 | 21 | 0.0%
Sema.NumDeclsDeserialized | 11,273,150 | 11,241,676 | -31,474 | -0.28%
Sema.NumLazyIterableDeclContexts | 1,411,874 | 1,400,218 | -11,656 | -0.83%
Sema.NumTypesDeserialized | 5,464,641 | 5,443,099 | -21,542 | -0.39%
Sema.NumTypesValidated | 497,726 | 497,726 | 0 | 0.0%
</details>